Moving item listeners out of actor-sheet
